# Optimize CPU load while Battlefield rendering