This PR relates to #7176 issue and optimizes the Battlefield render:
- render Battlefield background only when it is changed (grid on/off) and cache it;
- use `fheroes2::Copy()` instead of `fheroes2::Blit()` where possible:
  - when preparing the Battlefield background;
  - on every render when drawing the rectangle images without transparency (count background, status bar background, ...);
  - during DeathWave and EarthQuake spell effects.

Full render time for the Castle battlefield: 1.5 -> 0.85 ms
Full render time for the "Ship" battlefield, used in Battle only" mode: 1.0 -> 0.6 ms. 
<details><summary>Render time for optimized parts for the "Ship" battlefield</summary> 

```
RedrawCoverStatic render time:   0.2916 ms. -> 0.0443 ms.
RedrawPartialFinish render time: 0.1684 ms. -> 0.0142 ms.
RedrawInterface render time:     0.0159 ms. -> 0.0052 ms.
```

</details>