@Districh-ru 2023-05-21
Hi @kazenoyouni.
I made some minor optimizations in battle rendering. If you have time could you please check is the CPU load reduced in this PR on PlayStation Vita hardware. Here is the [compiled PS Vita binary.](https://github.com/ihhub/fheroes2/suites/13040703670/artifacts/706814327)