@kazenoyouni 2023-05-21
Hi @Districh-ru.
It works perfectly. CPU load reduced to 30%~40%!

https://github.com/ihhub/fheroes2/assets/111102725/b5246171-9b3d-46bf-ae5f-6198989e06d8