@Districh-ru 2023-05-22
> Hi @Districh-ru. It works perfectly. CPU load reduced to 30%~40%!

Great! @kazenoyouni, thank you for testing. If you find some other places where fheroes2 is heavily CPU intensive, please write.