@KOLANICH 2023-05-23
It increases memory footprint (while, I guess, the goal should be to decrease memory footprint, maybe even at cost of frame rendering time, it is a turn-based strategy, not a FPS), doesn't it? I wonder if it can be rerendered on each frames, but not by using software raster, but using OpenGL. Hexagons seem to be simple enough shapes to generate triangulation for them on fly and simple enough not to need a texture.