@Districh-ru 2023-05-24
> It increases memory footprint (while, I guess, the goal should be to decrease memory footprint, maybe even at cost of frame rendering time, it is a turn-based strategy, not a FPS), doesn't it? I wonder if it can be rerendered on each frames, but not by using software raster, but using OpenGL and rendering it on GPU. Hexagons seem to be simple enough shapes to generate triangulation for them on fly and simple enough not to need a texture.

Hi, @KOLANICH, yes this PR increases the memory footprint by 553.75 Kbytes. This is a cost for about 2 times less CPU usage. It is important for the handheld devices as it reduces the battery usage as long as for devices with low CPU (some old devices).
This PR does not only cache the hexagon grig. It aldo caches: ground, objects at the right and left parts of battlefield, big and small obstacles. These parts are never changed (except the rare switch of grid render).