@ihhub 2023-05-24
Hi @KOLANICH , we can't simply use any rendering technology out of the box as fheroes2 contains its own rendering engine and we call SDL (which is essentially a wrapper over DirectX or OpenGL) just to render already prepared frame. It is possible to use GPU memory and update data there but it requires a lot of code restructure and also data transfer between CPU and GPU is a bottleneck for most of operations on GPU especially when GPU doesn't support more than one DMA engines.

Talking about memory footprint: 0.5 MB is nothing in terms of memory. If we need we can just deallocate certain ICN images if we are hitting memory limit.