@KOLANICH 2023-05-24
>0.5 MB is nothing in terms of memory

It is a lot in terms of memory for such a device as Sony PSP. Which has 32 MiB of RAM, while for the original game in minimal system requirements it is written 8 MiB (and I guess it is not for the game alone, but for the typical system whole, a game developer cannot task consumers with measuring free memory on their systems, but everyone used to know total physical memory on their systems) and 16 MiB as recommended. According to #3112 this game alone requires 60 MiB of RAM, I guess on a 64-bit machine, then on 32-bit one it should require ~30% less. Also originally it didn't require a 3D-capable GPU, but PSP has a 3D-capable GPU with 2 MiB of RAM. I guess that PSP is more powerful than the computers the original game was intended to run on ... and still we cannot even think about porting this game to PSP before memory issue is resolved.

This is where the mindset "memory is not scarce, let's waste it!" (instead of the mindset "memory is always scarce", which is more aligned to reality) has lead us and the whole software industry (to the point that a modern browser cannot handle 10 tabs on a 2 GiB RAM 64-bit system and Xorg session is OOM-crashes (with taking all the active programs with itself) even on 8 GiB system, let alone 2 GiB one where it happens even more frequently (and happenned during I was writing this message) ) to.