@ihhub 2023-05-24
Another area where memory optimization and CPU is easy to do is **ImageRestorer** especially when it is called for a single-layered image. We can store just half of the image if possible.