@ihhub 2023-05-24
@Districh-ru and @KOLANICH , this pull request increases memory usage not by 553.75 Kbytes but by 519,680 bytes which is 507.5 KB. The main surface before the changes was 640 x 480 pixels while right now it is 640 x 443 pixels. However, we have 2 surfaces with double layer memory allocation.