@Districh-ru 2023-05-24
> @KOLANICH and @Districh-ru , if memory usage is a concern we can modify the whole rendering engine to properly handle single-layers images and do not store double memory for them as we do. In this case we can release a lot of memory, not only 0.5 MB for this case. The proposed approach uses a single layer image but only half of the allocated memory is utilized.

I already started this modification and will open a PR soon.